core: introduce event registry
The event registry is a storage of named trigger lists, called events.
One can find an event by its name, touch it (find or create), apply a
callback to all the events, and try to delete it - it will be actually
deleted only when its trigger list is empty.

Part of #8656

NO_CHANGELOG=internal
NO_DOC=internal