core: introduce trigger stable list
Since we are going to intoroduce trigger registry with method pairs, we
will need to implement stable iterator for lua triggers. But, actually,
it is already implemented under the hood - before trigger_run, all the
triggers from list are added to so-called run list which is stable.
Let's rename it to trigger stable list and put necessary methods to
the header.

Part of #8656

NO_CHANGELOG=internal
NO_DOC=internal