trigger: introduce persistent triggers
After this patch, functions with non-empty option `trigger` will be
inserted into the trigger registry. The option specifies event names in
which the function will be inserted as a trigger.

When a function is created, it is set as a trigger without any checks -
it can replace an existing one. When it is deleted, it can delete a
trigger it hasn't set - for example, it can happen when user manually
replaces a persistent trigger. So, it is recommended to have different
names (or even separate namespaces) for persistent and usual triggers.

In order not to create a new func adapter, all persistent triggers are
set to the trigger registry as Lua functions so we can still rely on
the fact that all the functions in the trigger registry are from Lua.

Note that both persistent and non-persistent funcs can be used as
triggers. Also, access rights are still being checked, so the triggers
will fail if a callee has no access to the function.

NO_CHANGELOG=later
NO_DOC=later

Closes #8663