trigger: introduce tarantool.trigger.on_change triggers
New event named 'tarantool.trigger.on_change' is called when any event
is modified (trigger.set or trigger.del). All the handlers are called with
one argument - name of the modified event. Returned value of each handler
is ignored. Handlers are fired after the event is changed (the event
contains inserted trigger, if any, and does not contain deleted one, if
any). In the case when an error is thrown trigger execution is
stopped, all the changes are rolled back and all called triggers are fired
again.

Speaking of the corner case when tarantool.trigger.on_change is modified,
new trigger, if any, is fired, and deleted trigger, if any, is not fired.
If the change is rolled back, all the called triggers, including a new one,
which was removed on rollback, are called in the same order. Old trigger,
which was not called on modification but was restored on rollback, is not
called as well.

Closes #8664

NO_CHANGELOG=later
NO_DOC=later