test: fix test case name duplication in iproto override test
Test box-luatest/iproto_request_handlers_overriding_test.lua has two
test cases with the same name. That is why a test case that was declared
first is replaced by the second one and that's why it is not launched.
The commit renames one of the cases.

NO_CHANGELOG=test
NO_DOC=test