core: introduce event_on_change function
Since we are going to move iproto override mechanism to the trigger
registry, we need to allow to set core triggers to on change event
because we need to notify iproto threades when a handler is overriden.
Let's add a new method that accepts a core trigger and sets it to an
internal trigger list.

NO_CHANGELOG=internal
NO_DOC=internal