port: introduce port_light
The patch introduces a new lightweight port with limited capacity which
is used to call triggers from tarantool. It can be dumped as many times
as you need.

It is called light because it stores only fundamental types and pointers,
it doesn't copy strings or MsgPack packets, so their lifetime is
completely defined by user. However, it references underlying tuples -
it is cheap anyway.

NO_TEST=later
NO_CHANGELOG=internal
NO_DOC=internal