box: persist gc consumers of non-anonymous replicas
The patch makes Tarantool persist gc consumers of replicas. Now every
creation and update of gc_consumer is a replace in space _gc_consumers
except for two cases - when schema does not support persistent consumers
(on downgrade) and when replica is joining - we have no entry in space
_cluster, so we cannot create a consumer - it will become hanging if the
join will fail, so the persistent one is created only when initial join
is completed. Note that when relay updates consumer for replica, the
error is only logged in the case of fail. It doesn't violate safety - we
will hold more xlogs than we need, but all needed xlogs won't be pruned.

It's important to underline that autorebootstrap is completely broken by
this commit - if replica has at least one relay, it has non-zero local
lsn and in this case autoreboostrap won't happen because do not remove
local data automatically.

In order to correctly upgrade to 3.2.0 in cluster, we set a persistent
on_replace trigger on space _index when creating _gc_consumers to persist
all existing gc consumers on all replicas.

Consumers are now dropped not only on deletion in _gc_consumers but also
in `replica_delete` function for two reasons. Firstly, the deletion can
be committed after the replica was deleted, so we will have no access to
the `replica->gc`, so destructor of `replica` has to unregister it.
Secondly, we can simply drop _gc_consumers space on downgrade, and all
the consumers will be still registered in replicas and will be correctly
unregistered on replica deletion.

Also, commit slightly reworks gc_consumer methods. Firstly,
gc_consumer_advance is removed because it is not needed anymore. Even
before the patch we had a place where the consumer was not advanced but
re-created because it could be moved back. Now every update of consumer
is a replace in _gc_consumers, and we don't know if it's supposed to be
advanced or not, so we simply re-create the consumer there. Secondly,
when registering and outdated consumer, it will become inactive and
will have no effect on cleanup.

Note that the solution does not work for anonymous replicas - we can't
store their consumers right in `replica->gc` because they are not added
to `replicaset` after restart. To solve this problem, we could store all
gc consumers in a special tree indexed by `uuid` or we could create
anonymous replicas on insert to `_gc_consumers`, but we will deal with
it in patch introducing gc consumers for anonymous replicas.

Along the way, the new error `ER_INVALID_VCLOCK` is introduced - it is
needed to raise an error when an MP_MAP representing vlock on replace in
_gc_consumers is invalid.

@TarantoolBot document
Title: Persistent WAL GC state

Now all replicas has WAL GC consumers persisted in space _gc_consumers.
When the replica is deleted from space _cluster, its consumer is deleted
as well. It ensures that xlogs needed for replica won't be deleted after
restart of instance. Note that they still can be deleted when we have no
space for newer xlogs.