schema: introduce space _gc_consumers
This space will be used to store gc consumers persistently. Since the
feature will be exported to 2.11.4 and we have linear version history,
this space is created on upgrade to 2.11.4, is dropped on upgrade to
3.0.0 and is created again on upgrade to 3.2.0. Using this approach,
the space has only two states - space is absent and space is created and
is consistent with space _cluster. We could leave the space on downgrade
from 3.2.0 to a version from [3.0.0, 3.2.0), but then we would care
about the cases when the space is created but is not supposed to be used
and when the space is created but it has stale consumers - this approach
is more complicated, so we decided to stick with the simple one.

The commit is dedicated to schema update, actual logic (on_replace trigger)
is implemented in the next commit.

Note that the new system space is treated in a special way - we don't
set our trigger for system spaces to it because it shouldn't acquire ddl
and limit depth of on_replace triggers. If we acquired ddl on replace in
_gc_consumers, user's transactions would be aborted for no apparent
reason when the consumers are advanced (that's what "acquire ddl"
actually do). And, since replace in `_gc_consumers` can be done from
on_replace trigger of `_cluster`, limiting depth of its triggers would
lead to error in such cases.

NO_CHANGELOG=later
NO_DOC=later