schema: introduce space _gc_consumers
This space will be used to store gc_consumers persistently. Since the
feature will be exported to 2.11.4 and we have linear version history,
this space is created on upgrade to 2.11.4 and on upgrade to 3.2.0 if
it does not exist.

When creating the space, we set a persistent on_replace trigger on
space _index to persist all existing gc consumers on all replicas.

The commit is dedicated to schema update, actual logic (on_replace trigger)
is implemented in the next commit.

Note that the new system space is treated in a special way - we don't
set our trigger for system spaces to it because it shouldn't acquire ddl
and limit depth of on_replace triggers.

NO_CHANGELOG=later
NO_DOC=later