box: deprecate wal_cleanup_delay
The commit deprecates option wal_cleanup_delay - it was introduced as a
workaround for a problem, which is solved by persistent gc state.

However, we automatically set wal_cleanup_delay to non-zero value if
current schema does not support _gc_consumers space and the instance
is a part of a replicaset. It prevents xlogs from being cleaned up on
upgrade to version with deprecated but still supported
wal_cleanup_delay option.

Part of tarantool/tarantool-ee#742

@TarantoolBot document
Title: Deprecate wal_cleanup_delay

The option `wal_cleanup_delay` is deprecated starting from Tarantool 3.2.0.
The new compat module option `box_cfg_wal_cleanup_delay` was
introduced to control whether the option is still available.
With the old behavior, which is the default in Tarantool 3.2.0,
the option is still available and has the same default value, but a
deprecation warning is logged when non-zero or non-default value is used.
With the new behavior, its default value is zero and an error is raised
when a non-zero value is used.

(Please create https://tarantool.io/compat/box_cfg_wal_cleanup_delay)

Note that the compat option can be set only before initial `box.cfg` call.

We are planning to switch the compat option to the new behavior
starting from Tarantool 4.0 with the ability to revert to the
old behavior. Starting from Tarantool 5.0 we are planning to
drop `wal_cleanup_delay` completely.