gc: persist gc consumers
The commit turns in-memory gc consumers into persistent ones.
Replicas' descriptors (`struct replica`) do not manage consumers
anymore - they are registered and unregistered on replace in space
`_cluster` instead. Since modification of persistent consumer is
more "heavy" operation (it can yield and block current fiber) the
commit introduces two types of update - synchronous and asynchronous.
The first one is needed, for example, on subscribe - we want to update
the consumer and only then send xlogs. The second one is needed for
relay not to litter tx fiber pool.

Note that the new system space is treated in a special way - we don't
set our trigger for system spaces to it because it shouldn't acquire ddl
and limit depth of on_replace triggers. If we acquired ddl on replace in
_gc_consumers, user's transactions would be aborted for no apparent
reason when the consumers are advanced (that's what "acquire ddl"
actually do). And, since replace in `_gc_consumers` can be done from
on_replace trigger of `_cluster`, limiting depth of its triggers would
lead to error in such cases. Also, we cannot use surrogate space for
`_gc_consumers` to set its on_replace trigger because it can be deleted
on upgrade to `3.0.0` and then the trigger will be dropped along with
the space. Fortunately, we don't need a surrogate space anyway.

Also, triggers for space `_gc_consumers` are placed not in `alter.cc`
but in `gc.c` to incapsulate internal methods.

About implementation of asynchronous updates: there were two
approaches - start a fiber for each relay that will wake up and
advance the consumer, or start one fiber in gc that will traverse
all consumers and persist "dirty" ones. The first soultion seems
to be easier - you don't need to think about synchronization of
synchronous and asynchronous updates of consumers, but the second
one feels more appropriate because even after relay is stopped
WAL GC fiber will advance consumer of disconnected replica if it
has any pending updates, and having one fiber seems more convenient
too. So the second solution was chosen. To easily implement
synchronous modifications then, we could simply use asynchronous
approach with `fiber_cond`, but it doesn't work in our case: we
register and unregister consumers inside on_replace trigger of
`_cluster` and therefore inside active transaction, so we cannot
yield there to wait for asynchronous update to complete.

Since there is no need to wait limbo to update the local space, let's
use flag TXN_FORCE_ASYNC when possible. Registration and unregistration
functions are called only from on_replace trigger of `_cluster`, hence,
inside active transactions, so they never use this flag - we mustn't set
this flag to synchronous transactions.

Also, in order to successfully deactivate existing consumer, a separate
method is needed because when consumer advanced, it is marked and when
persist fiber wakes up, it persists all such marked consumers, so a
simple replace to dummy consumer would lead to update by persist fiber
because it is still marked as asynchronously advanced - the special
method resets this flag.

Part of tarantool/tarantool-ee#742

@TarantoolBot document
Title: Persistent WAL GC state

Now all replicas has WAL GC consumers persisted in space _gc_consumers.
When the replica is deleted from space _cluster, its consumer is deleted
as well. It ensures that xlogs needed for replica won't be deleted after
restart of instance. Note that they still can be deleted when we have no
space for newer xlogs.