gc: slightly rework inactive consumers
Currently, every registered consumer is active by default (even if its
vclock is far behind from WAL GC one), consumer becomes inactive only
on ENOSPC and never becomes active again, which is not great - replica
could fill the xlog gap from another replica and continue to fetch our
xlogs but we won't pin them for it. In addition, when gc consumers will
be persistent, inactive consumers of hanging replicas will become active
after every restart. To relieve these problems, let's deactivate a
consumer on registration if it is outdated and activate it back when
it is advanced and becomes actual.