memtx: fix use-after-free on background index build
When building an index in background, we create on_rollback triggers for
tuples inserted concurrently. The problem here is on_rollback trigger
has independent from `index` and `memtx_ddl_state` lifetime - it can be
called after the index was build (and `memtx_ddl_state` is destroyed)
and even after the index was altered. So, in order to avoid
use-after-free in on_rollback trigger, let's not use `memtx_ddl_state`
there and store weak reference to the index.

Closes #10620

NO_DOC=bugfix