replication: store checkpoint reference in replica
On checkpoint join, we reference snapshot right before the join and
dereference it right after. The point of checkpoint join is that you can
try to fetch it again from the same snapshot on failure, so it makes
sense to dereference it only when the replica has completed the join or
has gone - that's what the commit does.

NO_TEST=in EE
NO_CHANGELOG=later
NO_DOC=later